+++ src/MSA/Identity.jl
@@ -16,12 +16,13 @@ function _percentidentity(seq1, seq2, len)
 end
 
 """
-Calculates the identity value between two aligned sequences.
+Calculates the fraction of identities between two aligned sequences.
 
 The identity value is calculated as the number of identical characters in the i-th position of both
 sequences divided by the length of both sequences.
 Positions with gaps in both sequences are not counted in the length of the sequence.
-Returns a value in [0, 1] range.
+Returns the fraction of identities, a value in the range [0, 1].
+One should multiply this fraction by 100 to get the percentage identity (PID).
 """
 function percentidentity(seq1, seq2)
     len = length(seq1)
@@ -33,7 +34,7 @@ end
 
 """
 Computes quickly if two aligned sequences have a identity value greater than a given `threshold` value.
-`threshold` should be a number in [0, 1] range.
+`threshold` should be a number in [0, 1] range (fraction of identities).
 Returns a boolean value.
 """
 function percentidentity(seq1, seq2, threshold::Float64)