+++ src/MSA/MultipleSequenceAlignment.jl
@@ -255,10 +255,10 @@ empty!(seq::AlignedSequence) = (empty!(seq.id); empty!(seq.index);
 # --------------------------
 
 """
-Calculates the percentage of gaps on the `Array` (alignment, sequence, column, etc.).
-This function can take an extra `dim` argument for calculation of the gap percentage over the given dimension
+Calculates the fraction of gaps on the `Array` (alignment, sequence, column, etc.).
+This function can take an extra `dim` argument for calculation of the gap fraction over the given dimension
 """
-function gappercentage(x::AbstractArray{Residue})
+function gapfraction(x::AbstractArray{Residue})
     counter = 0
     len = 0
     for res in x
@@ -268,14 +268,14 @@ function gappercentage(x::AbstractArray{Residue})
     float(counter) / float(len)
 end
 
-gappercentage(x::AbstractArray{Residue},
-              dim::Int) = vec( mapslices(gappercentage, x, dim) )
+gapfraction(x::AbstractArray{Residue},
+              dim::Int) = vec( mapslices(gapfraction, x, dim) )
 
 """
-Calculates the percentage of residues (no gaps) on the `Array` (alignment, sequence, column, etc.)
-This function can take an extra `dim` argument for calculation of the residue percentage over the given dimension
+Calculates the fraction of residues (no gaps) on the `Array` (alignment, sequence, column, etc.)
+This function can take an extra `dim` argument for calculation of the residue fraction over the given dimension
 """
-function residuepercentage(x::AbstractArray{Residue})
+function residuefraction(x::AbstractArray{Residue})
     counter = 0
     len = 0
     for res in x
@@ -285,20 +285,20 @@ function residuepercentage(x::AbstractArray{Residue})
     float(counter) / float(len)
 end
 
-residuepercentage(x::AbstractArray{Residue},
-                  dim::Int) = vec( mapslices(residuepercentage, x, dim) )
+residuefraction(x::AbstractArray{Residue},
+                  dim::Int) = vec( mapslices(residuefraction, x, dim) )
 
 """
 Coverage of the sequences with respect of the number of positions on the MSA
 """
-coverage(msa::Matrix{Residue}) = residuepercentage(msa, 2)
+coverage(msa::Matrix{Residue}) = residuefraction(msa, 2)
 coverage(msa::AbstractMultipleSequenceAlignment) = coverage(msa.msa)
 
 """
-Percentage of gaps per column/position on the MSA
+Fraction of gaps per column/position on the MSA
 """
-columngappercentage(msa::Matrix{Residue}) = gappercentage(msa, 1)
-columngappercentage(msa::AbstractMultipleSequenceAlignment) = columngappercentage(msa.msa)
+columngapfraction(msa::Matrix{Residue}) = gapfraction(msa, 1)
+columngapfraction(msa::AbstractMultipleSequenceAlignment) = columngapfraction(msa.msa)
 
 # Reference
 # ---------
@@ -364,7 +364,7 @@ function gapstrip!(msa::AbstractMultipleSequenceAlignment, annotate::Bool=true;
     if nsequences(msa) != 0
         annotate && annotate_modification!(msa, string("gapstrip! : Deletes columns with more than ",
                                                        gaplimit, " gaps."))
-        filtercolumns!(msa, columngappercentage(msa) .<= gaplimit, annotate)
+        filtercolumns!(msa, columngapfraction(msa) .<= gaplimit, annotate)
     else
         throw("There are not sequences in the MSA after coverage filter")
     end
@@ -388,7 +388,7 @@ function gapstrip(msa::Matrix{Residue}; coveragelimit::Float64=0.75,
         throw("There are not columns in the MSA after the gap trimming")
     end
     if nsequences(msa) != 0
-        msa = filtercolumns(msa, columngappercentage(msa) .<= gaplimit)
+        msa = filtercolumns(msa, columngapfraction(msa) .<= gaplimit)
     else
         throw("There are not sequences in the MSA after coverage filter")
     end